backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
bcrypt==3.2.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydub
speechrecognition
ffmpeg-python
pytest
httpx
pytest-cov
